Conflicts:
	cocos2d/CCAction.h
	cocos2d/CCActionInstant.h
	cocos2d/CCActionInstant.m
	cocos2d/CCActionInterval.m
	cocos2d/CCActionManager.h
	cocos2d/CCActionManager.m
	cocos2d/CCLabelAtlas.m
	cocos2d/ccConfig.h

+++ cocos2d/CCAtlasNode.h
@@ -36,20 +36,20 @@
  All features from CCNode are valid, plus the following features:
  - opacity and RGB colors
  */
-@interface CCAtlasNode : CCNode <CCRGBAProtocol, CCTextureProtocol> {
-
+@interface CCAtlasNode : CCNode <CCRGBAProtocol, CCTextureProtocol>
+{
 	// texture atlas
 	CCTextureAtlas	*textureAtlas_;
 
 	// chars per row
-	int				itemsPerRow_;
+	NSUInteger		itemsPerRow_;
 	// chars per column
-	int				itemsPerColumn_;
+	NSUInteger		itemsPerColumn_;
 		
 	// width of each char
-	int				itemWidth_;
+	NSUInteger		itemWidth_;
 	// height of each char
-	int				itemHeight_;
+	NSUInteger		itemHeight_;
 
 	// blend function
 	ccBlendFunc		blendFunc_;
@@ -74,10 +74,10 @@
 
 
 /** creates a CCAtlasNode  with an Atlas file the width and height of each item measured in points and the quantity of items to render*/
-+(id) atlasWithTileFile:(NSString*)tile tileWidth:(int)w tileHeight:(int)h itemsToRender: (int) c;
++(id) atlasWithTileFile:(NSString*)tile tileWidth:(NSUInteger)w tileHeight:(NSUInteger)h itemsToRender: (NSUInteger) c;
 
 /** initializes an CCAtlasNode  with an Atlas file the width and height of each item measured in points and the quantity of items to render*/
--(id) initWithTileFile:(NSString*)tile tileWidth:(int)w tileHeight:(int)h itemsToRender: (int) c;
+-(id) initWithTileFile:(NSString*)tile tileWidth:(NSUInteger)w tileHeight:(NSUInteger)h itemsToRender: (NSUInteger) c;
 
 /** updates the Atlas (indexed vertex array).
  * Shall be overriden in subclasses
